# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ethnos.app Academic Bibliography API",
    "version": "2.0.0",
    "description": "\nREST API for academic bibliography: search, works, persons, institutions, venues, courses, citations, collaborations and health.\n\nHighlights\n- Public endpoints; selected admin/metrics routes require 'x-access-key'.\n- Standard response envelope: { status, data, pagination?, meta? }.\n- Pagination supports page/limit and offset/limit.\n- Identifiers: DOI, ORCID, ROR, ISSN.\n\nNotes\n- Rate limiting is configurable via environment and enabled by default.\n- Performance and dataset size vary by deployment.\n        ",
    "contact": {
      "name": "Bruno Cesar Cunha Cruz, PhD Student"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-developer-orcid": "0000-0001-8652-2333",
    "x-institution": "PPGAS/MN/UFRJ"
  },
  "servers": [
    {
      "url": "https://api.ethnos.app",
      "description": "Production API Server - Enterprise-ready academic research infrastructure"
    },
    {
      "url": "http://localhost:1210",
      "description": "Development Server - Local testing and development environment"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "XAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-key",
        "description": "Internal access key required for protected endpoints (security, dashboard, health except /liveness)"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "Response payload. Varies per endpoint."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 6894,
            "description": "Total number of results"
          },
          "page": {
            "type": "integer",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "example": 20,
            "description": "Number of results per page"
          },
          "totalPages": {
            "type": "integer",
            "example": 345,
            "description": "Total number of pages"
          },
          "hasNext": {
            "type": "boolean",
            "example": true,
            "description": "Whether there is a next page"
          },
          "hasPrev": {
            "type": "boolean",
            "example": false,
            "description": "Whether there is a previous page"
          }
        }
      },
...
    },
    "parameters": {
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Number of items to return per page (works with both page and offset formats)",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "offsetParam": {
        "name": "offset",
        "in": "query",
        "description": "Number of items to skip. Automatically converts to page format: offset÷limit = page number",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        },
        "example": 20
      },
      "pageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based). Use with limit parameter for traditional pagination",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "example": 2
      },
...
    },
    "responses": {
      "Success": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SuccessEnvelope"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Resource not found",
              "code": "NOT_FOUND"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/subjects/{id}": {
      "get": {
        "summary": "Get subject by ID",
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/subjects/{id}/children": {
      "get": {
        "summary": "List subject children",
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
    "/subjects/{id}/hierarchy": {
      "get": {
        "summary": "Get subject hierarchy",
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/subjects/{id}/works": {
      "get": {
        "summary": "List works for a subject",
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "in": "query",
            "name": "min_relevance",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "year_from",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year_to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "document_type",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
    "/subjects/{id}/courses": {
      "get": {
        "summary": "List courses for a subject",
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "in": "query",
            "name": "year_from",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year_to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "program_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reading_type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
  }
}